started to get some of the actions in place .... i think
