+++ todo/.eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/bradleydion/Desktop/unit-3/sprint-2/week-1/reducer-todo/todo/src/index.js":"1","/Users/bradleydion/Desktop/unit-3/sprint-2/week-1/reducer-todo/todo/src/App.js":"2","/Users/bradleydion/Desktop/unit-3/sprint-2/week-1/reducer-todo/todo/src/reportWebVitals.js":"3","/Users/bradleydion/Desktop/unit-3/sprint-2/week-1/reducer-todo/todo/src/todo.js":"4","/Users/bradleydion/Desktop/unit-3/sprint-2/week-1/reducer-todo/todo/src/reducer/todoreducer.js":"5","/Users/bradleydion/Desktop/unit-3/sprint-2/week-1/reducer-todo/todo/src/action/todoaction.js":"6"},{"size":500,"mtime":1608156769836,"results":"7","hashOfConfig":"8"},{"size":156,"mtime":1608165010137,"results":"9","hashOfConfig":"8"},{"size":362,"mtime":1608156769837,"results":"10","hashOfConfig":"8"},{"size":870,"mtime":1608166160444,"results":"11","hashOfConfig":"8"},{"size":239,"mtime":1608167029756,"results":"12","hashOfConfig":"8"},{"size":236,"mtime":1608165180825,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"1553id3",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"23","usedDeprecatedRules":"16"},{"filePath":"24","messages":"25","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"26","messages":"27","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/bradleydion/Desktop/unit-3/sprint-2/week-1/reducer-todo/todo/src/index.js",[],["28","29"],"/Users/bradleydion/Desktop/unit-3/sprint-2/week-1/reducer-todo/todo/src/App.js",[],"/Users/bradleydion/Desktop/unit-3/sprint-2/week-1/reducer-todo/todo/src/reportWebVitals.js",[],"/Users/bradleydion/Desktop/unit-3/sprint-2/week-1/reducer-todo/todo/src/todo.js",["30"],"import React, {useState, useReducer} from \"react\"\nimport todoReducer from \"./reducer/todoreducer\"\n\nconst ToDo = () => {\n    const [newTask, setNewTask]=useState(\"\")\n    const incompleteTask = [\n        {\n        item: \"Learn About Reducers\",\n        completed:false,\n        id: Date.now()\n        }\n    ]\n    const [state, dispatch] = useReducer(todoReducer, incompleteTask)\n    const handleChanges = (e) =>{\n        setNewTask({item:e.target.value, completed:false, id:Date.now()})\n    }\n\n\n    return (\n        <div className= \"todo-list\">\n        <h1>Todo List</h1>\n        \n            <h2>To Be Done</h2>\n            <h3>{state.item}</h3>\n            <input \n                value = {newTask}\n                onChange ={handleChanges}\n                type =\"text\"\n                name =\"task\"\n                />\n        </div>\n        \n\n    )\n}\nexport default ToDo;","/Users/bradleydion/Desktop/unit-3/sprint-2/week-1/reducer-todo/todo/src/reducer/todoreducer.js",[],"/Users/bradleydion/Desktop/unit-3/sprint-2/week-1/reducer-todo/todo/src/action/todoaction.js",["31"],{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","replacedBy":"35"},{"ruleId":"36","severity":1,"message":"37","line":13,"column":19,"nodeType":"38","messageId":"39","endLine":13,"endColumn":27},{"ruleId":"40","severity":1,"message":"41","line":6,"column":1,"nodeType":"42","endLine":11,"endColumn":6},"no-native-reassign",["43"],"no-negated-in-lhs",["44"],"no-unused-vars","'dispatch' is assigned a value but never used.","Identifier","unusedVar","import/no-anonymous-default-export","Assign object to a variable before exporting as module default","ExportDefaultDeclaration","no-global-assign","no-unsafe-negation"]
+[{"/Users/bradleydion/Desktop/reducer-todo/todo/src/index.js":"1","/Users/bradleydion/Desktop/reducer-todo/todo/src/reportWebVitals.js":"2","/Users/bradleydion/Desktop/reducer-todo/todo/src/App.js":"3","/Users/bradleydion/Desktop/reducer-todo/todo/src/todo.js":"4","/Users/bradleydion/Desktop/reducer-todo/todo/src/toDoReducer.js":"5","/Users/bradleydion/Desktop/reducer-todo/todo/src/toDoAction.js":"6"},{"size":500,"mtime":1608156769000,"results":"7","hashOfConfig":"8"},{"size":362,"mtime":1608156769000,"results":"9","hashOfConfig":"8"},{"size":156,"mtime":1608165010000,"results":"10","hashOfConfig":"8"},{"size":960,"mtime":1608168063951,"results":"11","hashOfConfig":"8"},{"size":354,"mtime":1608168998777,"results":"12","hashOfConfig":"8"},{"size":316,"mtime":1608169018585,"results":"13","hashOfConfig":"8"},{"filePath":"14","messages":"15","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},"1p2ohdq",{"filePath":"17","messages":"18","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"16"},{"filePath":"21","messages":"22","errorCount":0,"warningCount":2,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":null},"/Users/bradleydion/Desktop/reducer-todo/todo/src/index.js",[],["27","28"],"/Users/bradleydion/Desktop/reducer-todo/todo/src/reportWebVitals.js",[],"/Users/bradleydion/Desktop/reducer-todo/todo/src/App.js",[],"/Users/bradleydion/Desktop/reducer-todo/todo/src/todo.js",["29","30"],"/Users/bradleydion/Desktop/reducer-todo/todo/src/toDoReducer.js",[],"/Users/bradleydion/Desktop/reducer-todo/todo/src/toDoAction.js",["31"],{"ruleId":"32","replacedBy":"33"},{"ruleId":"34","replacedBy":"35"},{"ruleId":"36","severity":1,"message":"37","line":13,"column":12,"nodeType":"38","messageId":"39","endLine":13,"endColumn":17},{"ruleId":"36","severity":1,"message":"40","line":13,"column":19,"nodeType":"38","messageId":"39","endLine":13,"endColumn":27},{"ruleId":"41","severity":1,"message":"42","line":6,"column":1,"nodeType":"43","endLine":14,"endColumn":6},"no-native-reassign",["44"],"no-negated-in-lhs",["45"],"no-unused-vars","'state' is assigned a value but never used.","Identifier","unusedVar","'dispatch' is assigned a value but never used.","import/no-anonymous-default-export","Assign object to a variable before exporting as module default","ExportDefaultDeclaration","no-global-assign","no-unsafe-negation"]
